Create Apps Script Code
This is the Apps Script code of the functions required to implement the import of files saved in Google Drive. It executes entirely in Google's server using  the Apps Script platform. It provides an ordered flow execution to download files located in multiple folders; ensure that no-duplicate files are downloaded; that the URL location of the original file is replaced by a local path so that domain users know where to locate the file. The execution finishes with the deletion of the downloaded files from Drive storage; enabling the preservation of available cloud storage capacity.